backend/requirements.txt
# ShapeBot Backend Requirements

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# CORS & Security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Redis (Session & Rate Limiting)
redis==5.0.1
aioredis==2.0.1

# Vector DB (Pinecone)
pinecone-client==3.0.0

# Logging & Monitoring
python-json-logger==2.0.7

# Environment Management
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
